Add gradle and change package name
