+++ .gitignore
@@ -1,0 +1,5 @@
+/.idea
+/.gradle/
+/gradle/
+
+/build
