--- BlackJack/src/contents/CPUTest.java
@@ -1,14 +1,0 @@
-package contents;
-
-import static org.junit.Assert.*;
-
-import org.junit.Test;
-
-public class CPUTest {
-
-	@Test
-	public void test() {
-		fail("まだ実装されていません");
-	}
-
-}
