+++ src/main/java/app/CPUTest.java
@@ -1,0 +1,14 @@
+//package contents;
+//
+//import static org.junit.Assert.*;
+//
+//import org.junit.Test;
+//
+//public class CPUTest {
+//
+//	@Test
+//	public void test() {
+//		fail("まだ実装されていません");
+//	}
+//
+//}
